refactor: 모임 참여, 모임 참여 취소 로직 hook으로 분리
