+++ .gitignore
@@ -39,3 +39,4 @@ next-env.d.ts
 .env.sentry-build-plugin
 
 *storybook.log
+.vscode/settings.json
