+++ .vscode/settings.json
@@ -1,23 +1,14 @@
 {
-  "cSpell.words": ["autodocs", "lucide", "selectbox", "signup", "testid"],
-  "workbench.colorCustomizations": {
-    "activityBar.activeBackground": "#4b2374",
-    "activityBar.background": "#4b2374",
-    "activityBar.foreground": "#e7e7e7",
-    "activityBar.inactiveForeground": "#e7e7e799",
-    "activityBarBadge.background": "#935f2c",
-    "activityBarBadge.foreground": "#e7e7e7",
-    "commandCenter.border": "#e7e7e799",
-    "panel.border": "#4b2374",
-    "sash.hoverBorder": "#4b2374",
-    "statusBar.background": "#32174d",
-    "statusBar.foreground": "#e7e7e7",
-    "statusBarItem.hoverBackground": "#4b2374",
-    "statusBarItem.remoteBackground": "#32174d",
-    "statusBarItem.remoteForeground": "#e7e7e7",
-    "titleBar.activeBackground": "#32174d",
-    "titleBar.activeForeground": "#e7e7e7",
-    "titleBar.inactiveBackground": "#32174d99",
-    "titleBar.inactiveForeground": "#e7e7e799"
-  }
+  "cSpell.words": [
+    "auths",
+    "autodocs",
+    "hookform",
+    "lucide",
+    "selectbox",
+    "Signin",
+    "signup",
+    "testid",
+    "usehooks",
+    "userid"
+  ]
 }
